@@ -1,10 +1,10 @@
 name: editable
 description: A highly customizable, editable data table (spreadsheet) package for Flutter projects.
-version: 1.1.4
+version: 2.0.0
 homepage: https://github.com/godilite/editable
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
